Adds documentation but with a bunch of leftover questions
After discovering some issues with the current code, it seemed best to
go through it from start to finish and document my understanding of
the code.  This has been a cursory reading with limited insights hence
there are a bunch of TODO's in the code requesting more information or
shedding light onto questions I have had.

We don't expect to answer all these questions soon.  They will be
removed as we continue to find better approaches on handling the
respective code.