Create documents before uploading attachments when processing deltas
When we receive a delta and it implies that we should create a new
document, we should ensure to create the document to which we will
upload the file before uploading the file itself.  Hence we move the
updating of the document's properties and the uploading of the
attachment to that specific field.  This seems to make the delta work
with changing data.