Experimentation with different types of search
We've added two experimental types of search (of which only one
works).  The fuzzy_phrase which should do phrase matching with fuzzy
terms and the fuzzy_terms which matches terms in a fuzzy way in the
requested field.  We could not get the former running but may get back
to that in the near future.