Correctly index documents with multiple values per field
Due to incorrect assumptions, fields with multiple values were not
being correctly indexed. This commit corrects that. There remains an
ambiguity between fields intended to be arrays; if these have only one
value, they will still be indexed as string values. This could be
defined more specifically in the config, if necessary.